fix: address review comments
* add docstring to properties and exported methods and class
* make mutable data structures private and expose
getter properties which return a copy of them
* move smithy projen source file to a new submodule
